+++ dx/src/exec/libdx/memorystubs.c
@@ -34,9 +34,9 @@ void _dxfcleanup_mem() { }
 
 int _dxf_initmemory() { return OK; }
 
-Error DXmemsize(int size) { return OK; }
+Error DXmemsize(uint64 size) { return OK; }
 
-Error DXSetMemorySize(unsigned int size, int ratio) { return OK; }
+Error DXSetMemorySize(uint64 size, int ratio) { return OK; }
 
 /* we can make this work for the pvs - there are #defines for it */
 Error DXGetMemorySize(unsigned int *sm, unsigned int *lg, unsigned int *lo)
